refactor: move AuthContext types into separate file
